update compress method API

+++ packages/polar/src/lib/deploy/compress.ts
@@ -2,7 +2,6 @@ import chalk from "chalk";
 import { execSync } from "child_process";
 import fs, { readdirSync } from "fs-extra";
 import path from "path";
-import zlib from "zlib";
 
 import {
   ARTIFACTS_DIR,
@@ -13,14 +12,17 @@ import {
   multiImageVersion,
   singleImageVersion
 } from "../../internal/core/project-structure";
-// eslint-disable-next-line
-import { cmpStr } from "../../internal/util/strings";
-import type { PolarRuntimeEnvironment } from "../../types";
 
 export async function compress (
-  networkName: string[],
-  env: PolarRuntimeEnvironment
+  contractName: string
 ): Promise<void> {
+  const sourcePath = path.join(CONTRACTS_DIR, "artifacts", `${contractName}.wasm`);
+  const destPath = path.join(CONTRACTS_OUT_DIR, `${contractName}_compressed.wasm`);
+
+  if (fs.existsSync(destPath)) {
+    console.log(`Compressed .wasm file exists for contract ${contractName}, skipping compression`);
+    return;
+  }
   const paths = readdirSync(CONTRACTS_DIR);
   if (paths.includes("Cargo.toml")) { // Only one contract in the contracts dir and compile in contracts dir only
     const dockerCmd = `docker run --rm -v ${path.resolve(CONTRACTS_DIR)}:/code \
@@ -31,13 +33,9 @@ export async function compress (
     console.log(chalk.greenBright(`Creating compressed .wasm file using cosmwasm/rust-optimizer:${singleImageVersion}...`));
     execSync(dockerCmd, { stdio: 'inherit' });
 
-    const fileContents = fs.createReadStream(path.join(CONTRACTS_DIR, "contract.wasm.gz"));
-    const writeStream = fs.createWriteStream(path.join(CONTRACTS_OUT_DIR, "contract.wasm"));
-    const unzip = zlib.createGunzip();
-
-    fileContents.pipe(unzip).pipe(writeStream);
-    fs.unlinkSync(path.join(CONTRACTS_DIR, "contract.wasm.gz"));
-    console.log(`Created file ${path.join(CONTRACTS_OUT_DIR, "contract.wasm")}`);
+    fs.copyFileSync(sourcePath, destPath);
+    fs.rmdirSync(path.join(CONTRACTS_DIR, "artifacts"), { recursive: true });
+    console.log(`Created file ${path.join(CONTRACTS_OUT_DIR, `${contractName}.wasm`)}`);
   } else { // Multiple contracts and each should be compiled by going inside each of them
     const dockerCmd = `docker run --rm -v ${path.resolve(CONTRACTS_DIR)}:/code \
                   --mount type=volume,source=${path.basename(ARTIFACTS_DIR)},target=/code/target \
